backend/requirements.txt
fastapi
uvicorn
python-dotenv
qdrant-client
tiktoken
requests
groq



